Post meta data syncs

+++ uninstall.php
@@ -29,13 +29,15 @@ delete_option('wawp_wa_credentials_valid');
 // Get array of restricted pages
 $restricted_pages = get_option('wawp_array_of_restricted_pages');
 // Loop through each page and delete our extra post meta
-if (isset($restricted_pages)) {
+if (isset($restricted_pages) && !empty($restricted_pages)) {
 	foreach ($restricted_pages as $restricted_page_id) {
 		delete_post_meta($restricted_page_id, 'wawp_restricted_groups');
 		delete_post_meta($restricted_page_id, 'wawp_restricted_levels');
 		delete_post_meta($restricted_page_id, 'wawp_is_page_restricted');
 	}
 }
+// delete_post_meta(2, 'wawp_groups');
+// delete_post_meta(2, 'wawp_levels');
 // Delete restricted pages option value
 delete_option('wawp_array_of_restricted_pages');
